refactor: plugins_page
