+++ packages/plugins/page/index.ts
@@ -11,8 +11,8 @@
  */
 
 import entry, { api } from './src/Main.vue'
-import metaData from './meta.js'
-import { PageService } from './src/composable'
+import metaData from './meta'
+import { PageService } from './src/composable/index'
 import PageGeneral from './src/PageGeneral.vue'
 import './src/styles/vars.less'
